refactor(ios): finish decomposing RootView into container + bootstrapper (#14348)
## Summary

Repo-quality loop, iteration 1 (no Linear issue — ad-hoc loop). Finishes the `RootView.swift` decomposition started in the iOS loop — the file was still **918 lines**, over the 500-line `file-size-check` bar. Now **405 lines**, and both new files are under the bar too.

- `LiveRootContainer` (the live Clerk auth container + deep-link/callback finalization) → new `App/LiveRootContainer.swift` (378 lines).
- The native auth plumbing it depends on → new `App/LiveAuthBootstrapper.swift` (142 lines): `LiveAuthBootstrapError`, `MobileAuthReturnError`, `MobileAuthFinalizationStageError`, `runMobileAuthFinalizationStage`, `LiveAuthBootstrapper`.
- Access changes: the three symbols `LiveRootContainer` references (`MobileAuthReturnError`, `runMobileAuthFinalizationStage`, `LiveAuthBootstrapper`) go `private` → internal; the two only used within the bootstrapper file stay `private`.
- `RootView.swift` keeps `AppContentView`, the `RootView` struct, and the `AppState` extension — verified to use none of the moved plumbing. Dropped the now-unused `ClerkKit` import.
- `project.pbxproj` wired with explicit entries (`plutil -lint` clean).

## Verification

- `pnpm run ios:lint` → **clean ✓**
- `pnpm run ios:test` (iPhone 17 sim, iOS 26.3.1) → **TEST SUCCEEDED** (full pass — this run also cleared the JOV-4252 flakes).

## Risks

- Low: verbatim same-module move, three access widenings, no logic changes. Auth-adjacent by file location only.
